Modifies Connection Behavior
This commit modifies MongoClient by renaming it to MongoDbClient and revises the connect method to pass the client db. This was done to enable an upgrade to the latest MongoDB node driver.

This commit also revises the coding style of each modified file to match project standards. As development moves forward and modified file should be revised to match these new standards.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>