Revises MongoDB Client
This commit revises the MongoDB client by removing the error check on the close method. The MongoGB Client will never throw an error. Its callback is always called with null.

This commit also revises the order of the connect function to throw an error before attempting to parse the client for a db.

Signed-off-by: Lui de la Parra <Lui@mutesymphony.com>